requirements.txt
faiss-cpu
transformers
sentence-transformers
langchain
tqdm
openai-whisper
pydub
bitsandbytes
pydantic
fastapi
chardet
langdetect
argostranslate
google-genai
requests
uvicorn
python-dotenv
Pillow
scipy
numpy
soundfile
langchain_community
google
google-generativeai
langchain_core
python-multipart
llama-cpp-python
